+++ alicloud/resource_alicloud_db_account_privilege.go
@@ -5,6 +5,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/aliyun/alibaba-cloud-sdk-go/services/rds"
+
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
@@ -55,13 +57,13 @@ func resourceAlicloudDBAccountPrivilege() *schema.Resource {
 
 func resourceAlicloudDBAccountPrivilegeCreate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*connectivity.AliyunClient)
-	rsdService := RdsService{client}
+	rdsService := RdsService{client}
 	instanceId := d.Get("instance_id").(string)
 	account := d.Get("account_name").(string)
 	privilege := d.Get("privilege").(string)
 	dbList := d.Get("db_names").(*schema.Set).List()
 	// wait instance running before granting
-	if err := rsdService.WaitForDBInstance(instanceId, Running, DefaultLongTimeout); err != nil {
+	if err := rdsService.WaitForDBInstance(instanceId, Running, DefaultLongTimeout); err != nil {
 		return WrapError(err)
 	}
 	d.SetId(fmt.Sprintf("%s%s%s%s%s", instanceId, COLON_SEPARATED, account, COLON_SEPARATED, privilege))
@@ -69,7 +71,7 @@ func resourceAlicloudDBAccountPrivilegeCreate(d *schema.ResourceData, meta inter
 	if len(dbList) > 0 {
 		for _, db := range dbList {
 			if err := resource.Retry(10*time.Minute, func() *resource.RetryError {
-				if err := rsdService.GrantAccountPrivilege(d.Id(), db.(string)); err != nil {
+				if err := rdsService.GrantAccountPrivilege(d.Id(), db.(string)); err != nil {
 					if IsExpectedErrors(err, OperationDeniedDBStatus) {
 						return resource.RetryableError(err)
 					}
@@ -110,6 +112,41 @@ func resourceAlicloudDBAccountPrivilegeRead(d *schema.ResourceData, meta interfa
 			names = append(names, pri.DBName)
 		}
 	}
+
+	if len(names) < 1 && strings.HasPrefix(object.DBInstanceId, "pgm-") {
+
+		request := rds.CreateDescribeDatabasesRequest()
+		request.RegionId = client.RegionId
+		request.DBInstanceId = object.DBInstanceId
+
+		err = resource.Retry(5*time.Minute, func() *resource.RetryError {
+			raw, err := client.WithRdsClient(func(rdsClient *rds.Client) (interface{}, error) {
+				return rdsClient.DescribeDatabases(request)
+			})
+			if err != nil {
+				if IsExpectedErrors(err, []string{"InternalError", "OperationDenied.DBInstanceStatus"}) {
+					return resource.RetryableError(WrapErrorf(err, DefaultErrorMsg, object.DBInstanceId, request.GetActionName(), AlibabaCloudSdkGoERROR))
+				}
+				return resource.NonRetryableError(WrapErrorf(err, DefaultErrorMsg, object.DBInstanceId, request.GetActionName(), AlibabaCloudSdkGoERROR))
+			}
+
+			addDebug(request.GetActionName(), raw, request.RpcRequest, request)
+
+			response, _ := raw.(*rds.DescribeDatabasesResponse)
+			for _, db := range response.Databases.Database {
+				for _, account := range db.Accounts.AccountPrivilegeInfo {
+					if account.Account == object.AccountName && (account.AccountPrivilege == parts[2] || account.AccountPrivilege == "ALL") {
+						names = append(names, db.DBName)
+					}
+				}
+			}
+			return nil
+		})
+		if err != nil {
+			return WrapError(err)
+		}
+	}
+
 	d.Set("db_names", names)
 
 	return nil
@@ -130,6 +167,9 @@ func resourceAlicloudDBAccountPrivilegeUpdate(d *schema.ResourceData, meta inter
 		add := ns.Difference(os).List()
 
 		if len(remove) > 0 {
+			if strings.HasPrefix(d.Id(), "pgm-") {
+				return WrapError(fmt.Errorf("At present, the PostgreSql database does not support revoking the current privilege."))
+			}
 			// wait instance running before revoking
 			if err := rdsService.WaitForDBInstance(parts[0], Running, DefaultTimeoutMedium); err != nil {
 				return WrapError(err)
@@ -172,6 +212,9 @@ func resourceAlicloudDBAccountPrivilegeDelete(d *schema.ResourceData, meta inter
 			return nil
 		}
 		return WrapError(err)
+	}
+	if strings.HasPrefix(d.Id(), "pgm-") {
+		return nil
 	}
 	var dbName string